+++ source/download/js/parts/Scroller.js
@@ -149,7 +149,9 @@ extend(defaultOptions, {
 });
 
 /**
- * The Navigator class
+ * The Navigator class.
+ *
+ * @class
  * @param {Object} chart
  */
 function Navigator(chart) {
@@ -207,8 +209,11 @@ Navigator.prototype = {
 
 		// Place it
 		handles[index][scroller.rendered && !scroller.hasDragged ? 'animate' : 'attr']({
-			translateX: scroller.scrollerLeft + scroller.scrollbarHeight + parseInt(x, 10),
-			translateY: scroller.top + scroller.height / 2 - 8
+			translateX: Math.round(
+				scroller.scrollerLeft + scroller.scrollbarHeight +
+					parseInt(x, 10)
+			),
+			translateY: Math.round(scroller.top + scroller.height / 2 - 8)
 		});
 	},
 
@@ -848,7 +853,7 @@ Navigator.prototype = {
 		if (this.series) {
 			this.removeBaseSeriesEvents();
 			each(this.series, function (s) { 
-				s.remove();
+				s.destroy();
 			});
 		}