+++ source/download/js/parts/Tooltip.js
@@ -15,7 +15,6 @@ var dateFormat = H.dateFormat,
 	merge = H.merge,
 	pick = H.pick,
 	splat = H.splat,
-	stop = H.stop,
 	syncTimeout = H.syncTimeout,
 	timeUnits = H.timeUnits;
 /**
@@ -74,6 +73,55 @@ H.Tooltip.prototype = {
 		});
 	},
 
+	/*= if (!build.classic) { =*/
+	/**
+	 * In styled mode, apply the default filter for the tooltip drop-shadow. It
+	 * needs to have an id specific to the chart, otherwise there will be issues
+	 * when one tooltip adopts the filter of a different chart, specifically one
+	 * where the container is hidden.
+	 */
+	applyFilter: function () {
+		
+		var chart = this.chart;
+		chart.renderer.definition({
+			tagName: 'filter',
+			id: 'drop-shadow-' + chart.index,
+			opacity: 0.5,
+			children: [{
+				tagName: 'feGaussianBlur',
+				in: 'SourceAlpha',
+				stdDeviation: 1
+			}, {
+				tagName: 'feOffset',
+				dx: 1,
+				dy: 1
+			}, {
+				tagName: 'feComponentTransfer',
+				children: [{
+					tagName: 'feFuncA',
+					type: 'linear',
+					slope: 0.3
+				}]
+			}, {
+				tagName: 'feMerge',
+				children: [{
+					tagName: 'feMergeNode'
+				}, {
+					tagName: 'feMergeNode',
+					in: 'SourceGraphic'
+				}]
+			}]
+		});
+		chart.renderer.definition({
+			tagName: 'style',
+			textContent: '.highcharts-tooltip-' + chart.index + '{' +
+				'filter:url(#drop-shadow-' + chart.index + ')' +
+			'}'
+		});
+	},
+	/*= } =*/
+	
+
 	/**
 	 * Create the Tooltip label element if it doesn't exist, then return the
 	 * label.
@@ -115,6 +163,13 @@ H.Tooltip.prototype = {
 					.shadow(options.shadow);
 				/*= } =*/
 			}
+			
+			/*= if (!build.classic) { =*/
+			// Apply the drop-shadow filter
+			this.applyFilter();
+			this.label.addClass('highcharts-tooltip-' + this.chart.index);
+			/*= } =*/
+
 			this.label
 				.attr({
 					zIndex: 8
@@ -391,7 +446,7 @@ H.Tooltip.prototype = {
 	refresh: function (point, mouseEvent) {
 		var tooltip = this,
 			chart = tooltip.chart,
-			label = tooltip.getLabel(),
+			label,
 			options = tooltip.options,
 			x,
 			y,
@@ -453,9 +508,10 @@ H.Tooltip.prototype = {
 			this.hide();
 		} else {
 
+			label = tooltip.getLabel();
+
 			// show it
 			if (tooltip.isHidden) {
-				stop(label);
 				label.attr({
 					opacity: 1
 				}).show();
@@ -466,7 +522,7 @@ H.Tooltip.prototype = {
 				this.renderSplit(text, chart.hoverPoints);
 			} else {
 				label.attr({
-					text: text.join ? text.join('') : text
+					text: text && text.join ? text.join('') : text
 				});
 
 				// Set the stroke color of the box to reflect the point
@@ -523,7 +579,7 @@ H.Tooltip.prototype = {
 
 			// Store the tooltip referance on the series
 			if (!tt) {
-				owner.tt = tt = ren.label(null, null, null, point.isHeader && 'callout')
+				owner.tt = tt = ren.label(null, null, null, 'callout')
 					.addClass('highcharts-tooltip-box ' + colorClass)
 					.attr({
 						'padding': options.padding,
@@ -535,27 +591,15 @@ H.Tooltip.prototype = {
 						/*= } =*/
 					})
 					.add(tooltipLabel);
-
-				// Add a connector back to the point
-				if (point.series) {
-					tt.connector = ren.path()
-						.addClass('highcharts-tooltip-connector ' + colorClass)
-						/*= if (build.classic) { =*/
-						.attr({
-							'stroke-width': series.options.lineWidth || 2,
-							'stroke': point.color || series.color || '${palette.neutralColor60}'
-						})
-						/*= } =*/
-						// Add it inside the label group so we will get hide and
-						// destroy for free
-						.add(tt);
-				}
 			}
 
 			tt.isActive = true;
 			tt.attr({
 				text: str
 			});
+			/*= if (build.classic) { =*/
+			tt.css(options.style);
+			/*= } =*/
 
 			// Get X position now, so we can move all to the other side in case of overflow
 			bBox = tt.getBBox();
@@ -599,37 +643,20 @@ H.Tooltip.prototype = {
 		// Distribute and put in place
 		H.distribute(boxes, chart.plotHeight + headerHeight);
 		each(boxes, function (box) {
-			var point = box.point,
-				tt = box.tt,
-				attr;
+			var point = box.point;
 
 			// Put the label in place
-			attr = {
+			box.tt.attr({
 				visibility: box.pos === undefined ? 'hidden' : 'inherit',
-				x: (rightAligned || point.isHeader ? box.x : point.plotX + chart.plotLeft + pick(options.distance, 16)),
-				y: box.pos + chart.plotTop
-			};
-			if (point.isHeader) {
-				attr.anchorX = point.plotX + chart.plotLeft;
-				attr.anchorY = attr.y - 100;
-			}
-			tt.attr(attr);
-
-			// Draw the connector to the point
-			if (!point.isHeader) {
-				tt.connector.attr({
-					d: [
-						'M',
-						point.plotX + chart.plotLeft - attr.x,
-						point.plotY + point.series.yAxis.pos - attr.y,
-						'L',
-						(rightAligned ? -1 : 1) * pick(options.distance, 16) +
-							point.plotX + chart.plotLeft - attr.x,
-						box.pos + chart.plotTop + tt.getBBox().height / 2 -
-							attr.y
-					]
-				});
-			}
+				x: (rightAligned || point.isHeader ? 
+					box.x :
+					point.plotX + chart.plotLeft + pick(options.distance, 16)),
+				y: box.pos + chart.plotTop,
+				anchorX: point.plotX + chart.plotLeft,
+				anchorY: point.isHeader ?
+					box.pos + chart.plotTop - 15 :
+					point.plotY + chart.plotTop
+			});
 		});
 	},