+++ changelog.md
@@ -19,3 +19,4 @@
 ## 1.3.0 [WIP]
 - Fixed issue with validation of Github Webhook.
 - Fixed issue with custom body parser.
+- Updated source files used in download builder to v5.0.4.